+++ my-fullstack-app/src/main.js
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
 import './style.css'
-import App from './App.vue'
+import App from './app.vue' 
 import router from './router/index.js'
 
 const app = createApp(App)
